MyProject/requirements.txt
Django==2.2.12
djangorestframework==3.12.4
numpy==1.21.2
